# openapi/parametrizacao.openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Fragmento da API de parametrizações da Atuária Previdenciária",
    "version": "0.0.1"
  },
  "tags": [
    {
      "name": "Parametrizacao"
    }
  ],
  "paths": {
    "/api/avaliacoes/{avaliacaoId}/parametrizacoes": {
      "get": {
        "tags": [
          "Parametrizacao"
        ],
        "operationId": "listarParametrizacoesAtuariais",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "avaliacaoId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "minimum": 1
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Versioned parameterizations for one evaluation",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ResumoParametrizacaoAtuarial"
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Parametrizacao"
        ],
        "operationId": "criarParametrizacaoAtuarial",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "avaliacaoId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "minimum": 1
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CriarParametrizacaoAtuarial"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Created draft parameterization",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ParametrizacaoAtuarial"
                }
              }
            }
          }
        }
      }
    },
    "/api/parametrizacoes/{id}": {
      "get": {
        "tags": [
          "Parametrizacao"
        ],
        "operationId": "obterParametrizacaoAtuarial",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Complete parameterization",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ParametrizacaoAtuarial"
                }
              }
            }
          },
          "404": {
            "description": "Parametrização não encontrada"
          }
        }
      },
      "patch": {
        "tags": [
          "Parametrizacao"
        ],
        "operationId": "atualizarParametrizacaoAtuarial",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AtualizarParametrizacaoAtuarial"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Updated draft metadata",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ParametrizacaoAtuarial"
                }
              }
            }
          }
        }
      }
    },
    "/api/parametrizacoes/{id}/parametros": {
      "patch": {
        "tags": [
          "Parametrizacao"
        ],
        "operationId": "definirParametrosAtuariais",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DefinirParametrosAtuariais"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Updated parameter values",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ParametrizacaoAtuarial"
                }
              }
            }
          }
        }
      }
    },
    "/api/parametrizacoes/{id}/candidato-aderencia": {
      "post": {
        "tags": [
          "Parametrizacao"
        ],
        "operationId": "promoverCandidatoAderencia",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PromoverCandidatoAderencia"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Candidate promoted into snapshot",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ParametrizacaoAtuarial"
                }
              }
            }
          }
        }
      }
    },
    "/api/parametrizacoes/{id}/hipotese/remover": {
      "post": {
        "tags": [
          "Parametrizacao"
        ],
        "operationId": "removerHipoteseAtuarial",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RemoverSelecaoHipoteseAtuarial"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Hypothesis removed from draft snapshot",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ParametrizacaoAtuarial"
                }
              }
            }
          }
        }
      }
    },
    "/api/parametrizacoes/{id}/aprovar": {
      "post": {
        "tags": [
          "Parametrizacao"
        ],
        "operationId": "aprovarParametrizacaoAtuarial",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Approved immutable snapshot",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ParametrizacaoAtuarial"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "DefinirValorParametroAtuarial": {
        "type": "object",
        "required": [
          "codigo",
          "categoria",
          "rotulo",
          "tipoValor",
          "jsonValor"
        ],
        "properties": {
          "tipoValor": {
            "type": "string",
            "enum": [
              "NUMBER",
              "INTEGER",
              "TEXT",
              "BOOLEAN"
            ]
          },
          "unidade": {
            "type": [
              "string",
              "null"
            ],
            "maxLength": 60
          },
          "codigo": {
            "type": "string",
            "minLength": 1,
            "maxLength": 120
          },
          "jsonValor": {
            "type": "string",
            "minLength": 1
          },
          "origem": {
            "type": [
              "string",
              "null"
            ],
            "maxLength": 120
          },
          "categoria": {
            "type": "string",
            "minLength": 1,
            "maxLength": 120
          },
          "rotulo": {
            "type": "string",
            "minLength": 1,
            "maxLength": 200
          }
        }
      },
      "SelecaoHipoteseAtuarial": {
        "type": "object",
        "required": [
          "id",
          "tipoHipotese",
          "estudoAderenciaId",
          "resultadoCandidatoId",
          "versaoBiometriaId",
          "codigoTabua",
          "nomeTabua",
          "rotuloVersao",
          "posicaoCandidata",
          "selecionadoEm"
        ],
        "properties": {
          "versaoBiometriaId": {
            "type": "string",
            "format": "uuid"
          },
          "resultadoCandidatoId": {
            "type": "string",
            "format": "uuid"
          },
          "selecionadoEm": {
            "type": "string",
            "format": "date-time"
          },
          "nomeTabua": {
            "type": "string"
          },
          "codigoTabua": {
            "type": "string"
          },
          "rotuloVersao": {
            "type": "string"
          },
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "tipoHipotese": {
            "type": "string"
          },
          "estudoAderenciaId": {
            "type": "string",
            "format": "uuid"
          },
          "posicaoCandidata": {
            "type": "integer",
            "minimum": 1
          }
        }
      },
      "AtualizarParametrizacaoAtuarial": {
        "type": "object",
        "properties": {
          "nome": {
            "type": "string",
            "minLength": 1,
            "maxLength": 200
          },
          "observacoes": {
            "type": [
              "string",
              "null"
            ],
            "maxLength": 4000
          }
        }
      },
      "RemoverSelecaoHipoteseAtuarial": {
        "type": "object",
        "required": [
          "selecaoId"
        ],
        "properties": {
          "selecaoId": {
            "type": "string",
            "format": "uuid"
          }
        }
      },
      "ValorParametroAtuarial": {
        "type": "object",
        "required": [
          "id",
          "codigo",
          "categoria",
          "rotulo",
          "tipoValor",
          "jsonValor",
          "unidade",
          "origem",
          "atualizadoEm"
        ],
        "properties": {
          "tipoValor": {
            "type": "string",
            "enum": [
              "NUMBER",
              "INTEGER",
              "TEXT",
              "BOOLEAN"
            ]
          },
          "unidade": {
            "type": [
              "string",
              "null"
            ]
          },
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "atualizadoEm": {
            "type": "string",
            "format": "date-time"
          },
          "codigo": {
            "type": "string"
          },
          "jsonValor": {
            "type": "string"
          },
          "origem": {
            "type": "string"
          },
          "categoria": {
            "type": "string"
          },
          "rotulo": {
            "type": "string"
          }
        }
      },
      "ParametrizacaoAtuarial": {
        "type": "object",
        "required": [
          "id",
          "avaliacaoId",
          "versao",
          "nome",
          "situacao",
          "observacoes",
          "criadoEm",
          "atualizadoEm",
          "aprovadoEm",
          "parametros",
          "hipoteses"
        ],
        "properties": {
          "criadoEm": {
            "type": "string",
            "format": "date-time"
          },
          "avaliacaoId": {
            "type": "integer",
            "minimum": 1
          },
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "atualizadoEm": {
            "type": "string",
            "format": "date-time"
          },
          "versao": {
            "type": "integer",
            "minimum": 1
          },
          "nome": {
            "type": "string"
          },
          "aprovadoEm": {
            "type": [
              "string",
              "null"
            ],
            "format": "date-time"
          },
          "observacoes": {
            "type": [
              "string",
              "null"
            ]
          },
          "hipoteses": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/SelecaoHipoteseAtuarial"
            }
          },
          "situacao": {
            "type": "string",
            "enum": [
              "RASCUNHO",
              "APROVADO",
              "SUBSTITUIDO"
            ]
          },
          "parametros": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ValorParametroAtuarial"
            }
          }
        }
      },
      "CriarParametrizacaoAtuarial": {
        "type": "object",
        "properties": {
          "nome": {
            "type": "string",
            "minLength": 1,
            "maxLength": 200
          },
          "copiarDeId": {
            "type": "string",
            "format": "uuid"
          },
          "observacoes": {
            "type": [
              "string",
              "null"
            ],
            "maxLength": 4000
          }
        }
      },
      "PromoverCandidatoAderencia": {
        "type": "object",
        "required": [
          "resultadoCandidatoId"
        ],
        "properties": {
          "resultadoCandidatoId": {
            "type": "string",
            "format": "uuid"
          }
        }
      },
      "DefinirParametrosAtuariais": {
        "type": "object",
        "required": [
          "parametros"
        ],
        "properties": {
          "parametros": {
            "type": "array",
            "minItems": 1,
            "items": {
              "$ref": "#/components/schemas/DefinirValorParametroAtuarial"
            }
          }
        }
      },
      "ResumoParametrizacaoAtuarial": {
        "type": "object",
        "required": [
          "id",
          "avaliacaoId",
          "versao",
          "nome",
          "situacao",
          "observacoes",
          "criadoEm",
          "atualizadoEm",
          "aprovadoEm"
        ],
        "properties": {
          "criadoEm": {
            "type": "string",
            "format": "date-time"
          },
          "avaliacaoId": {
            "type": "integer",
            "minimum": 1
          },
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "atualizadoEm": {
            "type": "string",
            "format": "date-time"
          },
          "versao": {
            "type": "integer",
            "minimum": 1
          },
          "nome": {
            "type": "string"
          },
          "aprovadoEm": {
            "type": [
              "string",
              "null"
            ],
            "format": "date-time"
          },
          "observacoes": {
            "type": [
              "string",
              "null"
            ]
          },
          "situacao": {
            "type": "string",
            "enum": [
              "RASCUNHO",
              "APROVADO",
              "SUBSTITUIDO"
            ]
          }
        }
      }
    }
  }
}